feat: showing the ball and the basket rim
