+++ src/css/index.css
@@ -2,6 +2,7 @@ body {
   font-family: "Berkelium", Arial, Helvetica, sans-serif;
   margin: 0;
   padding: 0;
+  background: #000;
 }
 
 span {
