+++ src/objects/Ball.ts
@@ -3,7 +3,8 @@ import { COLOR_PALETTE, GRID_UNIT, GROUND } from "~/GameConfig";
 export default class BallImageObject extends Phaser.GameObjects.Image {
   constructor(scene: Phaser.Scene, x, y, texture) {
     super(scene, x, y, texture);
-    this.setDisplaySize(GRID_UNIT, GRID_UNIT);
+    this.setDisplaySize(GRID_UNIT * 2, GRID_UNIT * 2);
+    this.setTintFill(COLOR_PALETTE.lightOrange);
     scene.children.add(this);
   }
