+++ src/scenes/GroundScene.ts
@@ -1,5 +1,5 @@
 import Phaser from "phaser";
-import { AnakondaObject } from "~/objects/Anakonda";
+import { BasketRimObject } from "~/objects/BasketRim";
 import { BorderObject } from "~/objects/Border";
 import BallImageObject from "~/objects/Ball";
 import GridSprite from "~/sprites/GridSprite";
@@ -23,7 +23,7 @@ export default class GroundScene extends Phaser.Scene {
     return new BallImageObject(this, x, y, texture);
   }
 
-  addPlayer(x = 0, y = 0): AnakondaObject {
-    return new AnakondaObject(this, x, y);
+  addBasketRim(x = 0, y = 0): BasketRimObject {
+    return new BasketRimObject(this, x, y);
   }
 }
