+++ src/scenes/LogoScene.ts
@@ -16,19 +16,19 @@ export default class GameScene extends Phaser.Scene {
     this.cameras.main.setBackgroundColor(COLOR_PALETTE.dark1);
     this.add.image(GRID_UNIT * 15, GRID_UNIT * 12, this.imageName).setScale(2);
 
-    this.cameras.main.fadeIn(500, 0, 0, 0);
-    this.time.delayedCall(1300, () => {
-      this.hideAndContinue(300);
+    this.cameras.main.fadeIn(100, 0, 0, 0);
+    this.time.delayedCall(100, () => {
+      this.hideAndContinue(100);
     });
 
     this.input.keyboard.once("keydown-SPACE", () => {
-      this.hideAndContinue(500);
+      this.hideAndContinue(0);
     });
 
     this.cameras.main.once(
       Phaser.Cameras.Scene2D.Events.FADE_OUT_COMPLETE,
       (camera, fx) => {
-        this.time.delayedCall(200, () => {
+        this.time.delayedCall(100, () => {
           this.scene.start("Loader", { fadeIn: true });
         });
       }
